raspiGrow/requirements.txt
CherryPy==4.0.0
DateTime==4.0.1
ephem==3.7.6.0
gevent==1.1rc3
greenlet==0.4.9
mysql-connector-python==2.0.3
pkginfo==1.2.1
requests==2.9.1
requests-toolbelt==0.6.0
signalr-client==0.0.6
simplejson==3.8.1
six==1.10.0
sseclient==0.0.11
twine==1.6.5
websocket-client==0.35.0
wheel==0.26.0
